requirements.txt
Flask==2.0.3
flask-cors==3.0.10
Werkzeug==2.0.3
torch==2.0.0
torchvision==0.15.0
Pillow==9.3.0
numpy>=1.24.3
albumentations==1.3.0
requests==2.31.0
python-dotenv==0.19.0